wasm + python examples
